Arreglo paquetes (#11)
* limpieza paquetes y test

* lo que era float cambio a double, correcciones errores de tipeo, nombres
de metodos viejos

+++ src/ar/edu/unq/po2/tpIntegrador/envio/RetiroEnSucursal.java
@@ -5,8 +5,8 @@ import ar.edu.unq.po2.tpIntegrador.state.Pedido;
 public class RetiroEnSucursal implements IFormaDeEnvio {
 
 	@Override
-	public Float calcularValorDelEnvio(Pedido pedido) {
-		return 0.0f;
+	public Double calcularValorDelEnvio(Pedido pedido) {
+		return 0.0;
 	}
 
 }
